tests/wpt/tests/css/css-images/parsing/image-resolution-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Images Module Level 4: parsing image-resolution with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-images-4/#propdef-image-resolution">
<meta name="assert" content="image-resolution supports only the grammar '[ from-image || <resolution> ] && snap?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("image-resolution", "auto");
test_invalid_value("image-resolution", "100%");
test_invalid_value("image-resolution", "2");
test_invalid_value("image-resolution", "3dpi snap from-image");
test_invalid_value("image-resolution", "from-image snap 4dppx");
</script>
</body>
</html>
